+++ js/inputmask.date.extensions.js
@@ -13,7 +13,7 @@ Optional extensions on the jquery.inputmask base
 		} else if (typeof exports === "object") {
 			module.exports = factory(require("./inputmask.dependencyLib.jquery"), require("./inputmask"));
 		} else {
-			factory(jQuery, window.Inputmask);
+			factory(window.dependencyLib || jQuery, window.Inputmask);
 		}
 	}
 	(function($, Inputmask) {